pyproject.toml
[project]
name = "matrix-controller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "dotenv>=0.9.9",
    "gspread-asyncio>=2.0.0",
    "pydantic>=2.11.9",
]
